DOMA-1012 move client fields to common fields
