+++ apps/condo/domains/meter/schema/MeterReading.js
@@ -2,16 +2,14 @@
  * Generated by `createschema meter.MeterReading 'number:Integer; date:DateTimeUtc; account?:Relationship:BillingAccount:SET_NULL; billingAccountMeter?:Relationship:BillingAccountMeter:SET_NULL; meter:Relationship:Meter:CASCADE; property:Relationship:Property:CASCADE; organization:Relationship:Organization:CASCADE; value:Integer; sectionName?:Text; floorName?:Text; unitName?:Text; client?:Relationship:User:SET_NULL; clientName?:Text; clientEmail?:Text; clientPhone?:Text; contact?:Relationship:Contact:SET_NULL; source:Relationship:MeterSource:PROTECT'`
  */
 
-const { Text, Relationship, Integer, DateTimeUtc } = require('@keystonejs/fields')
+const { Relationship, Integer, DateTimeUtc } = require('@keystonejs/fields')
 const { GQLListSchema } = require('@core/keystone/schema')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
-const { SENDER_FIELD, DV_FIELD } = require('@condo/domains/common/schema/fields')
+const { SENDER_FIELD, DV_FIELD, CONTACT_FIELD, CLIENT_EMAIL_FIELD, CLIENT_NAME_FIELD, CLIENT_PHONE_FIELD, CLIENT_FIELD } = require('@condo/domains/common/schema/fields')
 const access = require('@condo/domains/meter/access/MeterReading')
 const { getLastBillingAccountMeterReading } = require('../utils/serverSchema')
 const { VALUE_LESS_THAN_PREVIOUS_ERROR } = require('@condo/domains/meter/constants/errors')
 const { ORGANIZATION_OWNED_FIELD } = require('../../../schema/_common')
-const { PHONE_WRONG_FORMAT_ERROR } = require('@condo/domains/common/constants/errors')
-const { normalizePhone } = require('@condo/domains/common/utils/phone')
 
 
 const MeterReading = new GQLListSchema('MeterReading', {
@@ -102,48 +100,11 @@ const MeterReading = new GQLListSchema('MeterReading', {
             },
         },
 
-        client: {
-            schemaDoc: 'Inhabitant/customer/person who has a problem or want to improve/order something. Not null if we have a registered client',
-            type: Relationship,
-            ref: 'User',
-            kmigratorOptions: { null: true, on_delete: 'models.SET_NULL' },
-        },
-
-        contact: {
-            schemaDoc: 'Contact, that reported issue, described in this meter reading',
-            type: Relationship,
-            ref: 'Contact',
-            isRequired: false,
-            kmigratorOptions: { null: true, on_delete: 'models.SET_NULL' },
-        },
-
-        clientName: {
-            schemaDoc: 'Inhabitant/customer/person who has a problem. Sometimes we get a problem from an unregistered client, in such cases we have a null inside the `client` and just have something here. Or sometimes clients want to change it',
-            type: Text,
-        },
-
-        clientEmail: {
-            schemaDoc: 'Inhabitant/customer/person who has a problem. Sometimes we get a problem from an unregistered client, in such cases we have a null inside the `client` and just have something here. Or sometimes clients want to change it',
-            type: Text,
-        },
-
-        clientPhone: {
-            schemaDoc: 'Inhabitant/customer/person who has a problem. Sometimes we get a problem from an unregistered client, in such cases we have a null inside the `client` and just have something here. Or sometimes clients want to change it',
-            type: Text,
-            hooks: {
-                resolveInput: async ({ resolvedData }) => {
-                    if (!resolvedData['clientPhone']) return resolvedData['clientPhone']
-                    const newValue = normalizePhone(resolvedData['clientPhone'])
-                    return newValue || resolvedData['clientPhone']
-                },
-                validateInput: async ({ resolvedData, addFieldValidationError }) => {
-                    const newValue = normalizePhone(resolvedData['clientPhone'])
-                    if (resolvedData['clientPhone'] && newValue !== resolvedData['clientPhone']) {
-                        addFieldValidationError(`${PHONE_WRONG_FORMAT_ERROR}phone] invalid format`)
-                    }
-                },
-            },
-        },
+        client: CLIENT_FIELD,
+        contact: CONTACT_FIELD,
+        clientName: CLIENT_NAME_FIELD,
+        clientEmail:  CLIENT_EMAIL_FIELD,
+        clientPhone: CLIENT_PHONE_FIELD,
 
         source: {
             schemaDoc: 'Meter reading source channel/system. Examples: call, mobile_app, billing, ...',
